feat: Add subject and bio fields to User model
The code changes involve adding the "subject" and "bio" fields to the User model in the backend. These fields are required for creating a professor user. This update allows for storing additional information about professors in the database.

Please note that the commit message should not include any meta information such as issue references, tags, or author names.