feat: Update professorDetail.vue to retrieve user ID from JWT token
The code changes in professorDetail.vue involve updating the logic to retrieve the user ID from the JWT token stored in the localStorage. This allows for a more secure and reliable way of identifying the logged-in user. The decoded token is used to obtain the user ID, which is then used in creating a conversation between the user and the professor.

Please note that the commit message should not include any meta information such as issue references, tags, or author names.